requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
aiosqlite
redis
pydantic
pydantic-settings
python-dotenv
alembic
rich

# AI / RAG dependencies
pydantic-ai[google]
google-genai
pgvector

# Testing dependencies
pytest
pytest-asyncio
pytest-cov
httpx
